Move tombstone stuff into record_types.rs
