+++ sync15-adapter/src/record_types.rs
@@ -3,6 +3,85 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 // use error::{ErrorKind, Result};
+use bso_record::{BsoRecord, Sync15Record};
+
+#[derive(Deserialize, Serialize, Clone, Debug, PartialEq, Eq)]
+#[serde(untagged)]
+pub enum MaybeTombstone<T> {
+    Tombstone { id: String, deleted: bool },
+    Record(T)
+}
+
+impl<T> MaybeTombstone<T> {
+    #[inline]
+    pub fn tombstone<R: Into<String>>(id: R) -> MaybeTombstone<T> {
+        MaybeTombstone::Tombstone { id: id.into(), deleted: true }
+    }
+
+    #[inline]
+    pub fn is_tombstone(&self) -> bool {
+        match self {
+            &MaybeTombstone::Record(_) => false,
+            _ => true
+        }
+    }
+
+    #[inline]
+    pub fn unwrap(self) -> T {
+        match self {
+            MaybeTombstone::Record(record) => record,
+            _ => panic!("called `MaybeTombstone::unwrap()` on a Tombstone!"),
+        }
+    }
+
+    #[inline]
+    pub fn expect(self, msg: &str) -> T {
+        match self {
+            MaybeTombstone::Record(record) => record,
+            _ => panic!("{}", msg),
+        }
+    }
+
+    #[inline]
+    pub fn ok_or<E>(self, err: E) -> ::std::result::Result<T, E> {
+        match self {
+            MaybeTombstone::Record(record) => Ok(record),
+            _ => Err(err)
+        }
+    }
+
+    #[inline]
+    pub fn record(self) -> Option<T> {
+        match self {
+            MaybeTombstone::Record(record) => Some(record),
+            _ => None
+        }
+    }
+}
+
+impl<T> Sync15Record for MaybeTombstone<T> where T: Sync15Record {
+    fn collection_tag() -> &'static str { T::collection_tag() }
+    fn record_id(&self) -> &str {
+        match self {
+            &MaybeTombstone::Tombstone { ref id, .. } => id,
+            &MaybeTombstone::Record(ref record) => record.record_id()
+        }
+    }
+}
+
+impl<T> BsoRecord<MaybeTombstone<T>> {
+    #[inline]
+    pub fn is_tombstone(&self) -> bool {
+        self.payload.is_tombstone()
+    }
+
+    #[inline]
+    pub fn record(self) -> Option<BsoRecord<T>> where T: Clone {
+        self.map_payload(|payload| payload.record()).transpose()
+    }
+}
+
+pub type MaybeTombstoneRecord<T> = BsoRecord<MaybeTombstone<T>>;
 
 #[derive(Debug, Clone, Hash, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -38,4 +117,80 @@ pub struct PasswordRecord {
     pub times_used: Option<i64>,
 }
 
+impl Sync15Record for PasswordRecord {
+    fn collection_tag() -> &'static str { "passwords" }
+    fn record_id(&self) -> &str { &self.id }
+}
 
+#[cfg(test)]
+mod tests {
+
+    use super::*;
+    use key_bundle::KeyBundle;
+
+    #[derive(Serialize, Deserialize, Clone, PartialEq, Eq, Debug)]
+    struct DummyRecord {
+        id: String,
+        age: i64,
+        meta: String,
+    }
+
+    impl Sync15Record for DummyRecord {
+        fn collection_tag() -> &'static str { "dummy" }
+        fn record_id(&self) -> &str { &self.id }
+    }
+
+    #[test]
+    fn test_roundtrip_crypt_tombstone() {
+        let orig_record: MaybeTombstoneRecord<DummyRecord> = BsoRecord {
+            id: "aaaaaaaaaaaa".into(),
+            collection: "dummy".into(),
+            modified: 1234.0,
+            sortindex: None,
+            ttl: None,
+            payload: MaybeTombstone::tombstone("aaaaaaaaaaaa")
+        };
+
+        assert!(orig_record.is_tombstone());
+
+        let keybundle = KeyBundle::new_random().unwrap();
+
+        let encrypted = orig_record.clone().encrypt(&keybundle).unwrap();
+
+        assert!(keybundle.verify_hmac_string(
+            &encrypted.payload.hmac, &encrypted.payload.ciphertext).unwrap());
+
+        let decrypted: MaybeTombstoneRecord<DummyRecord> = encrypted.decrypt(&keybundle).unwrap();
+        assert!(decrypted.is_tombstone());
+        assert_eq!(decrypted, orig_record);
+    }
+
+    #[test]
+    fn test_roundtrip_crypt_record() {
+        let orig_record: MaybeTombstoneRecord<DummyRecord> = BsoRecord {
+            id: "aaaaaaaaaaaa".into(),
+            collection: "dummy".into(),
+            modified: 1234.0,
+            sortindex: None,
+            ttl: None,
+            payload: MaybeTombstone::Record(DummyRecord {
+                id: "aaaaaaaaaaaa".into(),
+                age: 105,
+                meta: "data".into()
+            })
+        };
+
+        assert!(!orig_record.is_tombstone());
+
+        let keybundle = KeyBundle::new_random().unwrap();
+
+        let encrypted = orig_record.clone().encrypt(&keybundle).unwrap();
+
+        assert!(keybundle.verify_hmac_string(
+            &encrypted.payload.hmac, &encrypted.payload.ciphertext).unwrap());
+
+        let decrypted: MaybeTombstoneRecord<DummyRecord> = encrypted.decrypt(&keybundle).unwrap();
+        assert!(!decrypted.is_tombstone());
+        assert_eq!(decrypted, orig_record);
+    }
+}
